Fixed Directory existence problem
the script is able to create subfolders if they do not exist.

+++ organise.py
@@ -1,20 +1,31 @@
 import os
 import shutil
-cwd=os.getcwd()
-print(cwd)
-print(os.path.isdir('.'))
-print(os.listdir(cwd))
 for(dirname,dirs,files) in os.walk('.'):
 	for filename in files:
 		if filename.endswith('.jpeg')|filename.endswith('.jpg')|filename.endswith('.png')|filename.endswith('.gif'):
-			shutil.move(filename,"Pictures")
+			if not os.path.exists("pictures"):
+				os.makedirs("pictures")
+			else:
+				shutil.move(filename,"pictures")
 		elif filename.endswith('.mp3')|filename.endswith('.wav')|filename.endswith('.WAV'):
-			shutil.move(filename,"music")
+			if not os.path.exists("music"):
+				os.makedirs("music")
+			else:
+				shutil.move(filename,"music")
 		elif filename.endswith('.mp4'):
-			shutil.move(filename,"video")
+			if not os.path.exists("video"):
+				os.makedirs("video")
+			else:
+				shutil.move(filename,"video")
 		elif filename.endswith('.avi'):
-			shutil.move(filename,'movies')
+			if not os.path.exists("movies"):
+				os.makedirs("movies")
+			else:
+				shutil.move(filename,'movies')
 		elif filename.endswith('.pdf')|filename.endswith('.doc')|filename.endswith('.docx')|filename.endswith('.mp4'):
-			shutil.move(filename,'documents')
+			if not os.path.exists("documents"):
+				os.makedirs("documents")
+			else:
+				shutil.move(filename,'documents')
